+++ src/queries/useBundleResultQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { getBundleResult } from "@/api/bundle/api";
+import { getBundleResult } from "@/api/my-bundle/api";
 
 export const useBundleResultQuery = (bundleId: string) => {
   return useQuery({
