+++ src/queries/useMyBundleDetailQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { getMyBundleDetail } from "@/api/bundle/api";
+import { getMyBundleDetail } from "@/api/my-bundle/api";
 
 export const useMyBundleDetailQuery = (id: number) => {
   return useQuery({
